[ICU-13725] Fix a few more MSVC warnings. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/41368 Former-commit-id: 9326b0c14788afa6ca799c1c241d5954916f5f24

+++ icu4c/source/test/intltest/numberformattesttuple.cpp
@@ -143,7 +143,7 @@ static void strToInt(
         status = U_ILLEGAL_ARGUMENT_ERROR;
         return;
     }
-    uint32_t value = 0;
+    int64_t value = 0;
     for (int32_t i = start; i < len; ++i) {
         UChar ch = str[i];
         if (ch < 0x30 || ch > 0x39) {
